+++ CHANGELOG.md
@@ -4,6 +4,14 @@ Mapbox welcomes participation and contributions from everyone.
 
 ## main
 
+### Features ✨ and improvements 🏁
+
+* Introduce new experimental `ViewAnnotation.enableSymbolLayerCollision` (`MapViewAnnotation.enableSymbolLayerCollision()` in SwiftUI) option which allows view annotations to hide underlying map symbols to avoid visual clutter.
+By default, the full bounding box of the view annotation is used for collision detection. If your annotation has a non-rectangular shape, it is highly recommended to mark the specific subviews that should participate via the new experimental `UIView.mbxCollisionBox` flag (`View.mbxCollisionBox()` in SwiftUI).
+* Support drawing view annotation collision boxes when `MapView.debugOptions` is set to `.collision`.
+* Improved `Marker` -- now it automatically hides the underlying map symbols below it's icon and text.
+* [SwiftUI] Added experimental `Map.viewAnnotationAvoidLayers()` which allows to set view annotations to avoid layers.
+
 ## 11.25.0 - 11 June, 2026
 
 ## 11.25.0-rc.2 - 04 June, 2026